Fix issue with with incorrect use of `a_start`
`shiny::shapes::bezier::intersections_in_range` was incorrectly using the `{a/b}_start` offsets when calculating the corresponding `{a/b}_end` value. This meant that end end values were almost always slightly off, thus requiring more iterations of the algorithm to arrive at an intersection.

This fix reduces the number of iterations to find an intersection for the included test from 11 to 8, a 27% improvement.